+++ src/main/java/aries/Aries.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class Aries {
     public static void main(String[] args) {
         Ui Ui = new Ui();
