+++ src/main/java/aries/AriesException.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class AriesException extends Exception {
     public AriesException(String message) {
         super(message);
