+++ src/main/java/aries/ByeCommand.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class ByeCommand implements Command {
     @Override
     public boolean execute(TaskList tasks, Ui ui) {
