+++ src/main/java/aries/Command.java
@@ -1,3 +1,5 @@
+package aries;
+
 public interface Command {
     boolean execute(TaskList tasks, Ui ui) throws AriesException;
