+++ src/main/java/aries/DeadlineCommand.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class DeadlineCommand implements Command {
     private String description;
