+++ src/main/java/aries/DeleteCommand.java
@@ -1,16 +1,18 @@
-public class MarkCommand implements Command {
+package aries;
+
+public class DeleteCommand implements Command {
     private String index;
 
-    public MarkCommand(String index) {
+    public DeleteCommand(String index) {
         this.index = index;
     }
 
     @Override
     public boolean execute(TaskList tasks, Ui ui) throws AriesException {
         int index = IndexHandling.getValidIndex(this.index, tasks.size());
-        Task taskToMark = tasks.get(index);
-        taskToMark.markAsDone();
-        ui.marked(taskToMark, true);
+        Task taskToDelete = tasks.get(index);
+        tasks.remove(index);
+        ui.deleted(taskToDelete, tasks);
         return true;
     }
 }
