+++ src/main/java/aries/Events.java
@@ -1,3 +1,5 @@
+package aries;
+
 import java.time.LocalDateTime;
 
 public class Events extends Task {
