+++ src/main/java/aries/ListCommand.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class ListCommand implements Command {
     @Override
     public boolean execute(TaskList tasks, Ui ui) {
