+++ src/main/java/aries/Storage.java
@@ -1,3 +1,5 @@
+package aries;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
