+++ src/main/java/aries/Task.java
@@ -1,3 +1,5 @@
+package aries;
+
 import java.io.Serializable;
 
 abstract class Task implements Serializable{
