+++ src/main/java/aries/ToDoCommand.java
@@ -1,3 +1,5 @@
+package aries;
+
 public class ToDoCommand implements Command {
     private String description;
